Move ssh-related code into separate package
Moved the HTTPOverSSHProxy struct into a separate package
